backend/requirements.txt
Django==3.0.4
django-allauth==0.41.0
django-coreapi==2.3.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
django-rest-knox==4.1.0
dnspython==1.16.0
docutils==0.15.2
entrypoints==0.3
eventlet==0.25.1
falcon==2.0.0
filelock==3.0.12
Flask==1.1.1
flask-restplus==0.13.0
gast==0.3.3

google-api-core==1.16.0
google-auth==1.13.1
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
greenlet==0.4.15
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.5.0
iplot==0.0.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lxml==4.2.1
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mongoengine==0.19.1
monotonic==1.5
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
nltk==3.4.5
notebook==6.0.3
numpy==1.18.2
oauth==1.0.1
oauthlib==3.1.0
pandas==1.0.2
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.4
protobuf==3.11.3
psutil==5.4.2
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
Pygments==2.6.1
pylint==2.4.4
pymongo==3.10.1
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python-debian==0.1.32
python3-openid==3.1.0
pytz==2019.3
pyxdg==0.25
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
reportlab==3.4.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.0
Rx==1.6.1
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
SecretStorage==2.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
simplejson==3.13.2
singledispatch==3.4.0.3
six==1.14.0
sklearn==0.0
smart-open==1.10.0
soupsieve==2.0
sqlparse==0.3.1
sympy==1.5.1

termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.21
wadllib==1.3.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.0.2
yellowbrick==1.1
zipp==3.1.0
zope.interface==4.3.2
